added middle repository on domain layer

+++ app/src/main/java/com/ilhomsoliev/todo/data/source/remote/models/request/TodoRequest.kt
@@ -1,10 +1,10 @@
 package com.ilhomsoliev.todo.data.source.remote.models.request
 
 data class TodoRequest(
-    val changed_at: Int,
+    val changed_at: Long,
     val color: String,
-    val created_at: Int,
-    val deadline: Int,
+    val created_at: Long,
+    val deadline: Long,
     val done: Boolean,
     val id: String,
     val importance: String,
